backend/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
psycopg2-binary==2.9.6
marshmallow-sqlalchemy==0.29.0
flask-marshmallow==0.14.0
flask-smorest==0.42.1
pyjwt==2.8.0
flask-bcrypt==1.0.1
marshmallow==3.20.1
flask_jwt_extended==4.5.2
flask-cors==3.0.10

torch==2.0.1
diffusers==0.12.1
invisible_watermark==0.2.0
transformers==4.27.4
accelerate==0.22.0
jax==0.4.1